+++ src/main/java/com/example/jpa/aula/infraestructure/repository/CozinhaRepositoryImpl.java
@@ -2,6 +2,7 @@ package com.example.jpa.aula.infraestructure.repository;
 
 import java.util.List;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -28,6 +29,7 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
         return entityManager.find(Cozinha.class, id);
     }
 
+    @Transactional
     @Override
     public Cozinha salvar(Cozinha cozinha) {
         return entityManager.merge(cozinha);
@@ -35,8 +37,13 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 
     @Override
     @Transactional
-    public void remover(Cozinha cozinha) {
-        cozinha = buscar(cozinha.getId());
+    public void remover(Long id) {
+        Cozinha cozinha = buscar(id);
+
+        if (cozinha == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
+
         entityManager.remove(cozinha);
     }
